chore: move js files to `lib/` fix: single entrypoint commander chaining refactor: --current-tag and --input-tag => --filter-tag doc: advanced example file with all options described
