+++ .prettierrc
@@ -1,6 +1,6 @@
 {
-    "trailingComma": "es5",
-    "singleQuote": true,
-    "printWidth": 100,
-    "semi": true
+  "trailingComma": "es5",
+  "singleQuote": true,
+  "printWidth": 100,
+  "semi": true
 }
