+++ examples/empty.js
@@ -1,0 +1,5 @@
+// easy-pingdom (c) 2019 Oak's Lab
+// This code is licensed under MIT license (see LICENSE for details)
+'use strict';
+
+// Nothing here intentionally
