+++ lib/pingdom.js
@@ -1,5 +1,6 @@
 // easy-pingdom (c) 2019 Oak's Lab
 // This code is licensed under MIT license (see LICENSE for details)
+'use strict';
 
 const fetch = require('node-fetch');
 const { URLSearchParams } = require('url');
@@ -14,7 +15,7 @@ const apiUrl = 'https://api.pingdom.com/api/2.1';
 
 const raw = (method, endpoint, params, body) =>
   fetch(
-    apiUrl + '/' + endpoint + (params ? '?' + params.toString() : ''),
+    apiUrl + '/' + endpoint + (params ? '?' + new URLSearchParams(params).toString() : ''),
     {
       method: method || 'get',
       headers,
