+++ lib/query.js
@@ -1,20 +1,23 @@
 // easy-pingdom (c) 2019 Oak's Lab
 // This code is licensed under MIT license (see LICENSE for details)
-const { readCurrent, logger, helpEnv, helpLogging } = require('../lib');
-const { readCheck } = require('../pingdom');
+'use strict';
+
+const { readCurrent, logger, helpEnv, helpLogging } = require('./lib');
+const { readCheck } = require('./pingdom');
 
 const queryCommand = program => {
-  program
-    .command('query')
+  const command = program.command('query');
+
+  command
     .usage('[options]')
     .option(
       '-f, --filter <regexp>',
-      'Expression for filtering input and current check names, ignore case',
+      'Expression for filtering input and current check names, ignore case'
     )
-    .option('    --current-tag <tag>', 'Tag name to be filtered on current')
+    .option('    --filter-tag <tag>', 'Include only checks with this tag')
     .option('-v, --verbose', 'Display detailed info about each check');
 
-  program.on('--help', () => {
+  command.on('--help', () => {
     console.log('');
 
     console.log('# Description');
@@ -26,12 +29,12 @@ const queryCommand = program => {
 
     console.log('# Examples:');
     console.log('## Display all current checks:');
-    console.log('./easy-pingdom/query.js | jq -C . | less -R');
+    console.log('./bin/easy-pingdom.js query | jq -C . | less -R');
   });
 
-  program.action(async ({ currentTag, filter, verbose }) => {
+  command.action(async ({ filterTag, filter, verbose }) => {
     try {
-      const current = await readCurrent({ currentTag, filter });
+      const current = await readCurrent({ filterTag, filter });
       if (!verbose) {
         console.log(JSON.stringify(Array.from(current.values())));
         return;
