+++ lib/update.js
@@ -1,18 +1,19 @@
 // easy-pingdom (c) 2019 Oak's Lab
 // This code is licensed under MIT license (see LICENSE for details)
+'use strict';
 
-const { helpEnv, helpLogging, logger, update, packageJson } = require('../lib');
+const { helpEnv, helpLogging, logger, update, packageJson } = require('./lib');
 
 const updateCommand = program => {
-  program
-    .command('update <inputFile>')
-    .usage('[options] <inputFile>')
+  const command = program.command('update <inputFile>');
+
+  command
+    .usage('[options] <input file>')
     .option(
       '-f, --filter <regexp>',
       'Expression for filtering input and current check names, ignore case',
     )
-    .option('    --input-tag <tag>', 'Tag name to be filtered on input')
-    .option('    --current-tag <tag>', 'Tag name to be filtered on current')
+    .option('    --filter-tag <tag>', 'Include only checks with this tag')
     .option('-U, --update', 'Create or update checks')
     .option('    --force-update', 'Update checks even without need to be updated')
     .option('-D, --delete', 'Delete all managed checks not included in input')
@@ -27,7 +28,7 @@ const updateCommand = program => {
       parseInt,
     );
 
-  program.on('--help', () => {
+  command.on('--help', () => {
     console.log('');
     console.log('# Description');
     console.log('Script takes input JSON file and updates current Pingdom configuration');
@@ -38,18 +39,18 @@ const updateCommand = program => {
 
     console.log('# Examples:');
     console.log('## Display what would be done:');
-    console.log('scripts/update.js --filter "example" --json example.js | jq -C . | less -R');
+    console.log('./bin/easy-pingdom.js update --filter "example" --json examples/simple.js | jq -C . | less -R');
     console.log('');
     console.log('## Update:');
-    console.log('scripts/update.js --filter "example" --update example.js');
+    console.log('./bin/easy-pingdom.js update --filter "example" --update examples/simple.js');
     console.log('');
     console.log('## Update everything:');
-    console.log('./easy-pingdom/update.js --update --delete --no-paused example.js');
+    console.log('./bin/easy-pingdom.js update --update --delete --no-paused examples/simple.js');
   });
 
-  const config = setupConfig(program);
+  command.action((inputFile, options) => {
+    // const config = setupConfig(program);
 
-  program.action((inputFile, options) => {
     update(inputFile, options)
       .then(() => logger.debug('SUCCESS'))
       .catch(err => {
